refactor(playerAnalysis): stand the standing on its own
The standing reads raw scores and renders before the analysis lands, so it is a
different input from the rest of the summary. Its describe block already used
none of the summary's fixtures.

isWeekOver walks every pick of every player twice, and the dialog body ran it
three times per render. The summary now works it out once and hands it to both
halves.
